Build 9 (FINAL)
This version dont work properly, so i will tracking the error an fix it in the future:

The color change after taking a point doesnt work

The audio doesnt pitch correctly when taking many points.
The text dont change the color properly when taking many points.
The UI is Bugged

When the game starts, the maxScore text will dissapear

The "TAP THE SCREEN TO PLAY" text will fading in and out

The buttons on the left or right ofthe screen will show the modal window to the center of the screen (BUGGED)